# Switching HGCal electrons and photons to TICL-based reconstruction