--- a/RecoEgamma/EgammaElectronAlgos/interface/GsfElectronAlgo.h
+++ b/RecoEgamma/EgammaElectronAlgos/interface/GsfElectronAlgo.h
@@ -80,12 +80,14 @@ public:
     bool pureTrackerDrivenEcalErrorFromSimpleParameterization;
     // ambiguity solving
     bool applyAmbResolution;              // if not true, ambiguity solving is not applied
+    bool ignoreNotPreselected;
     unsigned ambSortingStrategy;          // 0:isBetter, 1:isInnermost
     unsigned ambClustersOverlapStrategy;  // 0:sc adresses, 1:bc shared energy
     // for backward compatibility
     bool ctfTracksCheck;
     float PreSelectMVA;
     float MaxElePtForOnlyMVA;
+    bool useDefaultEnergyCorrection;
     // GED-Regression (ECAL and combination)
     bool useEcalRegression;
     bool useCombinationRegression;