--- a/RecoEgamma/EgammaElectronProducers/python/gsfElectrons_cfi.py
+++ b/RecoEgamma/EgammaElectronProducers/python/gsfElectrons_cfi.py
@@ -28,11 +28,8 @@ from Configuration.Eras.Modifier_pp_on_AA_2018_cff import pp_on_AA_2018
 pp_on_AA_2018.toModify(ecalDrivenGsfElectrons.preselection, minSCEtBarrel = 15.0)
 pp_on_AA_2018.toModify(ecalDrivenGsfElectrons.preselection, minSCEtEndcaps = 15.0)
 
-import FWCore.ParameterSet.Config as cms
-
 ecalDrivenGsfElectronsFromMultiCl = ecalDrivenGsfElectrons.clone(
     gsfElectronCoresTag = "ecalDrivenGsfElectronCoresFromMultiCl",
-    
     useGsfPfRecTracks = False,
     useDefaultEnergyCorrection = False,
     ambClustersOverlapStrategy = 0,