#### PR description:

1. HGCal electron (and photon) sequence modified to use TICL as input, instead of the old Multiclusters.
In addition, now there is an option to filter the input TICL tracksters based on their PID. This is turned off for now, but will be switched on soon in the future.
Files affected:
RecoParticleFlow/PFClusterProducer/python/particleFlowClusterHGC_cfi.py
RecoParticleFlow/PFClusterProducer/plugins/PFClusterFromHGCalMultiCluster.cc

2. Have turned off the old (default) track-based correction for electrons by adding a configurable boolean "useDefaultEnergyCorrection".
Files affected:
RecoEgamma/EgammaElectronProducers/plugins/GsfElectronProducer.cc
RecoEgamma/EgammaElectronAlgos/interface/GsfElectronAlgo.h
RecoEgamma/EgammaElectronAlgos/src/GsfElectronAlgo.cc

3. TICL-based electron customization (like ambiguity resolution).
Files affected:
RecoEgamma/EgammaElectronProducers/python/gsfElectrons_cfi.py


#### PR validation:

1. Test:
scram b distclean 
git cms-checkdeps -a -A
scram b -j 8
scram b runtests

Outcome: OK.

2. Test:
runTheMatrix.py -l limited -i all --ibeos

Outcome:
35 23 22 15 8 1 0 0 0 tests passed, 0 11 0 0 0 0 0 0 0 failed

3. Test:
scram build code-checks

Outcome: OK.

4. Test:
scram build code-format

Outcome: OK.