@cmsbuild 2020-09-21
A new Pull Request was created by  @SohamBhattacharya  for master.

It involves the following packages:

RecoEgamma/EgammaElectronAlgos
RecoEgamma/EgammaElectronProducers
RecoParticleFlow/PFClusterProducer


@perrotta, @jpata, @cmsbuild, @slava77 can you please review it and eventually sign? Thanks.
@mmarionncern, @Sam-Harper, @felicepantaleo, @riga, @afiqaize, @varuns23, @rovere, @lgray, @sobhatta, @clelange, @lecriste, @hatakeyamak, @jainshilpi, @seemasharmafnal, @cbernet this is something you requested to watch as well.
@silviodonato, @dpiparo, @qliphy you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>