@cmsbuild 2020-09-22
+1
Tested at: 7a741e305e3c01d6a4538fe5bab698601c6bac1a
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-15bf68/9478/summary.html
CMSSW: CMSSW_11_2_X_2020-09-21-2300
SCRAM_ARCH: slc7_amd64_gcc820