@cmsbuild 2020-09-22
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-15bf68/9478/summary.html

Comparison Summary:
 - No significant changes to the logs found
 - Reco comparison results: 871 differences found in the comparisons
 - DQMHistoTests: Total files compared: 35
 - DQMHistoTests: Total histograms compared: 2540471
 - DQMHistoTests: Total failures: 1339
 - DQMHistoTests: Total nulls: 0
 - DQMHistoTests: Total successes: 2539110
 - DQMHistoTests: Total skipped: 22
 - DQMHistoTests: Total Missing objects: 0
 - DQMHistoSizes: Histogram memory added: 0.0 KiB( 34 files compared)
 - Checked 149 log files, 22 edm output root files, 35 DQM output files