@SohamBhattacharya 2020-09-22
The comparison differences are in the HGCal e/gamma variables, as these objects are affected by the switch to TICL-based reconstruction.