@SohamBhattacharya 2020-09-22
> Thanks for the tests - I agree that the reco differences look fine in principle, being limited to HGCAL EGamma. I left some code suggestions.

Thanks for the suggestions. I have updated the repo with the changes.