@jpata 2020-09-25
With the last test, for completeness, I wanted to check the igprof timing [comparisons](https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-15bf68/9550/profiling/23434.21/RES_CPU_compare_23434.21.txt).

The modules that you modify do not seem to have a significant relative change, which is good:
- `PFClusterProducer::produce`: 3.1% -> 3.2% (55.48 s -> 61.3 s)
- `GsfElectronProducer::produce`: 0.6% -> 0.5% (9.99 s -> 10.44 s)

However, I noticed also a significant change in:
- `cms::CkfTrackCandidateMakerBase::produceBase`: 14.3% -> 16.3% (254.31 s -> 311.66 s)

Do you think this is relevant to your PR, or somehow spurious? I'm not familiar with the HGCAL reco chain (yet), so I don't know if modifying e.g. HGCAL clusters would change the inputs to `CkfTrackCandidateMakerBase`, most likely you already can answer that quickly 🙂.